Fix MSVC /W4 warnings in tutorials
Resolve all MSVC warnings under /W4 /WX across every tutorial and the
shared tutorial infrastructure.

Warning categories fixed:
  C4189  unused initialized local (alloc.h 'done' pattern)
  C4245  signed/unsigned mismatch — ray/shadow mask initialisation
         (mask = -1 -> 0xFFFFFFFFu) and size_t argument mismatches
  C4389  signed/unsigned mismatch in == operator
  C4456  local hides outer local — shadowed loop variables (i, j),
         mesh, hgeom, vertices, quad, tri
  C4457  local hides function parameter (v in geometry lambdas)
  C4458  local hides class member (N, ivariant, sampler)
  C4459  local hides global — renderPixelStandard/postIntersect/
         createPoints 'data' parameter renamed to 'td' throughout;
         quaternion_motion_blur 'qdc' parameter renamed
  C4505  unreferenced static function — benchmark.h helpers changed
         from 'static MAYBE_UNUSED' to 'inline'
  C5051  [[maybe_unused]] / [[fallthrough]] require /std:c++17;
         replaced with MAYBE_UNUSED macro and /* falls through */
         comments respectively

+++ tutorials/common/scenegraph/texture.cpp
@@ -19,7 +19,7 @@ namespace embree
   }
 
   Texture::Texture () 
-    : width(-1), height(-1), format(INVALID), bytesPerTexel(0), width_mask(0), height_mask(0), data(nullptr) {}
+    : width((unsigned int)-1), height((unsigned int)-1), format(INVALID), bytesPerTexel(0), width_mask(0), height_mask(0), data(nullptr) {}
   
   Texture::Texture(Ref<Image> img, const std::string &&fileName)
     : width(unsigned(img->width)), height(unsigned(img->height)), format(RGBA8), bytesPerTexel(4), width_mask(0), height_mask(0), data(nullptr), fileName(fileName)
